refactor: 아이콘 에셋 import 경로 정리
