--- src/assets/icons/index.ts
@@ -1,3 +1,0 @@
-export { default as GoogleIcon } from './google.svg'
-export { default as KakaoIcon } from './kakao.svg'
-export { default as NaverIcon } from './naver.svg'
