+++ src/pages/auth/components/FeatureVisual.tsx
@@ -1,9 +1,9 @@
 import { Clock4, MapPin, UserRound } from 'lucide-react'
 
+import Plus from '@/assets/icons/common/plus.svg?react'
 import LandingCalendar from '@/assets/login/LandingCalendar_noBack.svg?react'
 import LandingTodo from '@/assets/login/LandingTodo.svg?react'
 import { SparkleIcon } from '@/features/Home/Icon/SparkleIcon'
-import Plus from '@/shared/assets/icons/plus.svg?react'
 
 import * as S from '../Landing.styles'
