+++ src/shared/ui/calendar/CustomDatePicker/CustomDatePicker.tsx
@@ -1,7 +1,7 @@
 /** @jsxImportSource @emotion/react */
 import { useMemo, useState } from 'react'
 
-import Arrow from '@/shared/assets/icons/chevron.svg?react'
+import Arrow from '@/assets/icons/common/chevron.svg?react'
 import { theme } from '@/shared/styles/theme'
 import { type DatePickerRenderProps } from '@/shared/types/event/event'
 import { toDateValue } from '@/shared/utils/date'
