+++ src/shared/ui/common/Toast/Toast.tsx
@@ -2,10 +2,10 @@ import type { ComponentType, SVGProps } from 'react'
 import { useEffect, useState } from 'react'
 
 import CheckIcon from '@/assets/icons/check.svg?react'
+import CloseIcon from '@/assets/icons/common/close.svg?react'
 import WarningIcon from '@/assets/icons/exclamation.svg?react'
 import FolderIcon from '@/assets/icons/folder.svg?react'
 import InfoIcon from '@/assets/icons/uil_info.svg?react'
-import CloseIcon from '@/shared/assets/icons/close.svg?react'
 import type { ToastType } from '@/shared/types/common/toast'
 
 import { TOAST_VARIANTS } from './Toast.constants'
