+++ src/shared/ui/scheduleTodo/SearchPlace/SearchPlace.tsx
@@ -1,7 +1,7 @@
 import { Map, MapMarker, useKakaoLoader } from 'react-kakao-maps-sdk'
 
-import Check from '@/shared/assets/icons/check.svg?react'
-import Search from '@/shared/assets/icons/search.svg?react'
+import Check from '@/assets/icons/common/check.svg?react'
+import Search from '@/assets/icons/common/search.svg?react'
 import type { SearchPlaceProps } from '@/shared/types/schedule/types'
 import { DEFAULT_CENTER } from '@/shared/utils/searchPlace'
 import useSearchPlace from '@/shared/utils/useSearchPlace'
